refactor: 에러 상황 스낵바 duration Long으로 수정

+++ android/app/src/main/java/com/on/turip/ui/compose/login/LoginScreen.kt
@@ -61,7 +61,7 @@ fun LoginScreen(
                     val errorUiModel = effect.errorUiState.toUiModel() ?: return@collect
                     snackbarHostState.showSnackbar(
                         message = context.getString(errorUiModel.titleRes),
-                        duration = SnackbarDuration.Short,
+                        duration = SnackbarDuration.Long,
                     )
                 }
             }
